skip ssh host_key_checking
Rationale: this mechanism doesn't work well (ie: regularly fails to
fetch remote key, breaking provisioning).
And as it's based on running `ssh-keyscan` an attacker could simply do a
mitm attack at this moment. ie: there's no tier validation of the key
ansible trusts.
So basically, with or without this flag, it boils down to trusting the
IP address returned by the cloud-provider API is correct.